floatに対応した

+++ Assets/Script/Cat.cs
@@ -3,9 +3,9 @@ using UnityEngine;
 
 public class Cat : MonoBehaviour
 {
-    public string catName;           // ネコの名前
-    public int targetMeal;           // ネコの規定ごはん量
-    public SpriteRenderer catSprite; // ネコの見た目
+    public string catName;                      // ネコの名前
+    public float targetMeal;                    // ネコの規定ごはん量
+    public SpriteRenderer catSprite;            // ネコの見た目
     public TextMeshProUGUI catNameText;         // 名前を表示するUI（任意）
 
     // 表情差分（インスペクタでセット）
@@ -25,7 +25,7 @@ public class Cat : MonoBehaviour
     }
 
     // GameManagerから呼ばれる：データ設定
-    public void SetCatData(string name, int food)
+    public void SetCatData(string name, float food)
     {
         catName = name;
         targetMeal = food;
